refactor: 컨트롤러 책임 분리
